+++ src/main/java/com/zerobase/sns/domain/follow/entity/Follow.java
@@ -1,0 +1,41 @@
+package com.zerobase.sns.domain.follow.entity;
+
+import com.zerobase.sns.domain.user.entity.User;
+import com.zerobase.sns.global.entity.BaseEntity;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Entity
+public class Follow extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne
+  @JoinColumn(name = "follower_id", nullable = false)
+  private User follower;
+
+  @ManyToOne
+  @JoinColumn(name = "following_id", nullable = false)
+  private User following;
+
+  @Enumerated(EnumType.STRING)
+  private FollowStatus status = FollowStatus.DEFAULT;
+}
