+++ src/main/java/com/zerobase/sns/domain/likes/entity/Likes.java
@@ -1,7 +1,8 @@
-package com.zerobase.sns.domain.story.visitor;
+package com.zerobase.sns.domain.likes.entity;
 
-import com.zerobase.sns.domain.story.Story;
-import com.zerobase.sns.domain.user.User;
+import com.zerobase.sns.domain.post.entity.Post;
+import com.zerobase.sns.domain.user.entity.User;
+import com.zerobase.sns.global.entity.BaseEntity;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -18,17 +19,17 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class Visitor {
+public class Likes extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne
-  @JoinColumn(name = "story_id")
-  private Story story;
+  @JoinColumn(name = "user_id")
+  private User user;
 
   @ManyToOne
-  @JoinColumn(name = "visitor_id")
-  private User visitor;
+  @JoinColumn(name = "post_id")
+  private Post post;
 }
