+++ src/main/java/com/zerobase/sns/domain/message/entity/Message.java
@@ -1,10 +1,8 @@
-package com.zerobase.sns.domain.follow;
+package com.zerobase.sns.domain.message.entity;
 
-import com.zerobase.sns.domain.user.User;
+import com.zerobase.sns.domain.user.entity.User;
 import com.zerobase.sns.global.entity.BaseEntity;
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -14,28 +12,27 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
-@Setter
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class Follow extends BaseEntity {
+public class Message extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
   @ManyToOne
-  @JoinColumn(name = "follower_id", nullable = false)
-  private User follower;
+  @JoinColumn(name = "sender_id")
+  private User sender;
 
   @ManyToOne
-  @JoinColumn(name = "following_id", nullable = false)
-  private User following;
+  @JoinColumn(name = "receiver_id")
+  private User receiver;
 
-  @Enumerated(EnumType.STRING)
-  private FollowStatus status = FollowStatus.DEFAULT;
+  private String message;
+
+  private Boolean isDeleted;
 }
