+++ src/main/java/com/zerobase/sns/domain/reply/entity/Reply.java
@@ -1,7 +1,7 @@
-package com.zerobase.sns.domain.alarm;
+package com.zerobase.sns.domain.reply.entity;
 
-import com.zerobase.sns.domain.post.tag.Tag;
-import com.zerobase.sns.domain.user.User;
+import com.zerobase.sns.domain.comment.entity.Comment;
+import com.zerobase.sns.domain.user.entity.User;
 import com.zerobase.sns.global.entity.BaseEntity;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -9,7 +9,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,19 +19,19 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class Alarm extends BaseEntity {
+public class Reply extends BaseEntity {
 
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   private Long id;
 
-  private String content;
+  private String reply;
+
+  @ManyToOne
+  @JoinColumn(name = "comment_id")
+  private Comment parentComment;
 
   @ManyToOne
   @JoinColumn(name = "user_id")
   private User user;
-
-  @OneToOne
-  @JoinColumn(name = "tag_id")
-  private Tag tag;
 }
